# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Resposta detalhada da validação de e-mails",
        "title": "API de Validação de E-mails",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Marcos Oleniuk (Autor)",
            "email": "marcos@moleniuk.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/emails": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Valida uma lista de e-mails passados no corpo da requisição",
                "summary": "Valida múltiplos e-mails",
                "parameters": [
                    {
                        "description": "Lista de e-mails para validação",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.BulkEmailRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Chave de autenticação",
                        "name": "X-API-KEY",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.ValidationResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/valida-email": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Valida um e-mail passado via query string",
                "summary": "Valida um único e-mail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "E-mail para validação",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Chave de autenticação",
                        "name": "X-API-KEY",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.ValidationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.BulkEmailRequest": {
            "description": "Lista de e-mails para validação",
            "type": "object",
            "required": [
                "emails"
            ],
            "properties": {
                "emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "main.ValidationResponse": {
            "description": "Resposta detalhada da validação de e-mails",
            "type": "object",
            "properties": {
                "domain": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "is_valid": {
                    "type": "boolean"
                },
                "validation_type": {
                    "type": "string"
                }
            }
        }
    }
}